bibliotheque/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.14.0,<3.15.0
pytest>=7.4.0,<7.5.0
pytest-django>=4.5.2,<4.6.0
pytest-cov>=4.1.0,<4.2.0
django-filter>=23.5,<23.6
